refactor: Rename exceptions.build_env to build
